Add SettingsRepository
