--- app/src/main/java/com/brewthings/app/ui/screen/onboarding/OnboardingScreen.kt
@@ -1,7 +1,0 @@
-package com.brewthings.app.ui.screen.onboarding
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun OnboardingScreen() {
-}
