--- app/src/main/res/drawable/ic_device_unknown.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960">
-  <path
-      android:pathData="M480,680q-17,0 -29.5,-12.5T438,638q0,-17 12.5,-29.5T480,596q17,0 29.5,12.5T522,638q0,17 -12.5,29.5T480,680ZM450,552q0,-46 7.5,-63t42.5,-47q14,-14 24,-27.5t10,-30.5q0,-18 -13.5,-32T480,338q-27,0 -41,15.5T420,386l-54,-22q12,-35 41,-59.5t73,-24.5q47,0 80.5,25.5T594,382q0,24 -12,45t-30,39q-30,30 -36,42t-6,44h-60ZM280,920q-33,0 -56.5,-23.5T200,840v-720q0,-33 23.5,-56.5T280,40h400q33,0 56.5,23.5T760,120v720q0,33 -23.5,56.5T680,920L280,920ZM280,800v40h400v-40L280,800ZM280,720h400v-480L280,240v480ZM280,160h400v-40L280,120v40ZM280,160v-40,40ZM280,800v40,-40Z"
-      android:fillColor="#000000"/>
-</vector>
